+++ sdk/eventhubs/azure-messaging-eventhubs/src/main/java/com/azure/messaging/eventhubs/EventHubClientBuilder.java
@@ -92,6 +92,9 @@ import java.util.ServiceLoader;
 @ServiceClientBuilder(serviceClients = {EventHubProducerAsyncClient.class, EventHubProducerClient.class,
     EventHubConsumerAsyncClient.class, EventHubConsumerClient.class})
 public class EventHubClientBuilder {
+    // Default number of events to fetch when creating the consumer.
+    static final int DEFAULT_PREFETCH_COUNT = 500;
+
     /**
      * The name of the default consumer group in the Event Hubs service.
      */
@@ -105,19 +108,17 @@ public class EventHubClientBuilder {
      */
     private static final int MAXIMUM_PREFETCH_COUNT = 8000;
 
-    // Default number of events to fetch when creating the consumer.
-    static final int DEFAULT_PREFETCH_COUNT = 500;
     private static final String EVENTHUBS_PROPERTIES_FILE = "azure-messaging-eventhubs.properties";
     private static final String NAME_KEY = "name";
     private static final String VERSION_KEY = "version";
     private static final String UNKNOWN = "UNKNOWN";
 
-    private final ClientLogger logger = new ClientLogger(EventHubClientBuilder.class);
-
     private static final String AZURE_EVENT_HUBS_CONNECTION_STRING = "AZURE_EVENT_HUBS_CONNECTION_STRING";
     private static final AmqpRetryOptions DEFAULT_RETRY = new AmqpRetryOptions()
         .setTryTimeout(ClientConstants.OPERATION_TIMEOUT);
 
+    private final ClientLogger logger = new ClientLogger(EventHubClientBuilder.class);
+    private final Object connectionLock = new Object();
     private TokenCredential credentials;
     private Configuration configuration;
     private ProxyOptions proxyOptions;
@@ -130,6 +131,11 @@ public class EventHubClientBuilder {
     private EventHubConnectionProcessor eventHubConnectionProcessor;
     private int prefetchCount;
     private boolean isSharedConnection;
+
+    /**
+     * Keeps track of the open clients that were created from this builder when there is a shared connection.
+     */
+    private volatile int openClients;
 
     /**
      * Creates a new instance with the default transport {@link AmqpTransportType#AMQP} and a non-shared connection. A
@@ -240,7 +246,9 @@ public class EventHubClientBuilder {
      * @return The updated {@link EventHubClientBuilder} object.
      */
     public EventHubClientBuilder shareConnection() {
-        this.isSharedConnection = true;
+        synchronized (connectionLock) {
+            this.isSharedConnection = true;
+        }
         return this;
     }
 
@@ -466,17 +474,26 @@ public class EventHubClientBuilder {
 
         final MessageSerializer messageSerializer = new EventHubMessageSerializer();
 
-        if (isSharedConnection && eventHubConnectionProcessor == null) {
-            eventHubConnectionProcessor = buildConnectionProcessor(messageSerializer);
-        }
+        final EventHubConnectionProcessor processor;
+        synchronized (connectionLock) {
+            if (isSharedConnection) {
+                if (eventHubConnectionProcessor == null) {
+                    eventHubConnectionProcessor = buildConnectionProcessor(messageSerializer);
+                }
 
-        final EventHubConnectionProcessor processor = isSharedConnection
-            ? eventHubConnectionProcessor
-            : buildConnectionProcessor(messageSerializer);
+                final int numberOfOpenClients = ++openClients;
+                logger.info("# of open clients with shared connection: {}", numberOfOpenClients);
+            }
+
+            processor = isSharedConnection
+                ? eventHubConnectionProcessor
+                : buildConnectionProcessor(messageSerializer);
+        }
 
         final TracerProvider tracerProvider = new TracerProvider(ServiceLoader.load(Tracer.class));
 
-        return new EventHubAsyncClient(processor, tracerProvider, messageSerializer, scheduler, isSharedConnection);
+        return new EventHubAsyncClient(processor, tracerProvider, messageSerializer, scheduler, isSharedConnection,
+            this::onClientClose);
     }
 
     /**
@@ -510,6 +527,19 @@ public class EventHubClientBuilder {
         return new EventHubClient(client, retryOptions);
     }
 
+    void onClientClose() {
+        synchronized (connectionLock) {
+            final int numberOfOpenClients = --openClients;
+            logger.info("Closing a dependent client. # of open clients: {}", numberOfOpenClients);
+
+            if (numberOfOpenClients == 0) {
+                logger.info("No more open clients, closing shared connection.");
+                eventHubConnectionProcessor.dispose();
+                eventHubConnectionProcessor = null;
+            }
+        }
+    }
+
     private EventHubConnectionProcessor buildConnectionProcessor(MessageSerializer messageSerializer) {
         final ConnectionOptions connectionOptions = getConnectionOptions();
         final TokenManagerProvider tokenManagerProvider = new AzureTokenManagerProvider(
@@ -525,13 +555,13 @@ public class EventHubClientBuilder {
         final Flux<EventHubAmqpConnection> connectionFlux = Mono.fromCallable(() -> {
             final String connectionId = StringUtil.getRandomString("MF");
 
-            return (EventHubAmqpConnection) new EventHubReactorAmqpConnection(connectionId, connectionOptions, provider,
-                handlerProvider, tokenManagerProvider, messageSerializer, product, clientVersion);
+            return (EventHubAmqpConnection) new EventHubReactorAmqpConnection(connectionId, connectionOptions,
+                eventHubName, provider, handlerProvider, tokenManagerProvider, messageSerializer, product,
+                clientVersion);
         }).repeat();
 
         return connectionFlux.subscribeWith(new EventHubConnectionProcessor(
-            connectionOptions.getFullyQualifiedNamespace(), connectionOptions.getEntityPath(),
-            connectionOptions.getRetry()));
+            connectionOptions.getFullyQualifiedNamespace(), eventHubName, connectionOptions.getRetry()));
     }
 
     private ConnectionOptions getConnectionOptions() {
@@ -566,8 +596,8 @@ public class EventHubClientBuilder {
             ? CbsAuthorizationType.SHARED_ACCESS_SIGNATURE
             : CbsAuthorizationType.JSON_WEB_TOKEN;
 
-        return new ConnectionOptions(fullyQualifiedNamespace, eventHubName, credentials, authorizationType,
-            transport, retryOptions, proxyOptions, scheduler);
+        return new ConnectionOptions(fullyQualifiedNamespace, credentials, authorizationType, transport, retryOptions,
+            proxyOptions, scheduler);
     }
 
     private ProxyOptions getDefaultProxyConfiguration(Configuration configuration) {