+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/implementation/ServiceBusReactorAmqpConnection.java
@@ -8,19 +8,27 @@ import com.azure.core.amqp.AmqpRetryPolicy;
 import com.azure.core.amqp.AmqpSession;
 import com.azure.core.amqp.implementation.AmqpReceiveLink;
 import com.azure.core.amqp.implementation.AmqpSendLink;
+import com.azure.core.amqp.implementation.AzureTokenManagerProvider;
+import com.azure.core.amqp.implementation.CbsAuthorizationType;
 import com.azure.core.amqp.implementation.ConnectionOptions;
 import com.azure.core.amqp.implementation.MessageSerializer;
 import com.azure.core.amqp.implementation.ReactorConnection;
 import com.azure.core.amqp.implementation.ReactorHandlerProvider;
 import com.azure.core.amqp.implementation.ReactorProvider;
+import com.azure.core.amqp.implementation.RequestResponseChannel;
 import com.azure.core.amqp.implementation.RetryUtil;
+import com.azure.core.amqp.implementation.TokenManager;
 import com.azure.core.amqp.implementation.TokenManagerProvider;
 import com.azure.core.amqp.implementation.handler.SessionHandler;
+import com.azure.core.credential.TokenCredential;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.servicebus.models.ReceiveMode;
+import org.apache.qpid.proton.amqp.transport.ReceiverSettleMode;
+import org.apache.qpid.proton.amqp.transport.SenderSettleMode;
 import org.apache.qpid.proton.engine.BaseHandler;
 import org.apache.qpid.proton.engine.Session;
 import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Scheduler;
 
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -39,13 +47,16 @@ public class ServiceBusReactorAmqpConnection extends ReactorConnection implement
      * load balance messages is the eventHubName.
      */
     private final ConcurrentHashMap<String, AmqpSendLink> sendLinks = new ConcurrentHashMap<>();
-    private final Mono<ServiceBusManagementNode> managementChannelMono;
+    private final ConcurrentHashMap<String, ServiceBusManagementNode> managementNodes = new ConcurrentHashMap<>();
     private final String connectionId;
     private final ReactorProvider reactorProvider;
     private final ReactorHandlerProvider handlerProvider;
     private final TokenManagerProvider tokenManagerProvider;
     private final AmqpRetryOptions retryOptions;
     private final MessageSerializer messageSerializer;
+    private final TokenCredential tokenCredential;
+    private final Scheduler scheduler;
+    private final String fullyQualifiedNamespace;
 
     /**
      * Creates a new AMQP connection that uses proton-j.
@@ -62,31 +73,57 @@ public class ServiceBusReactorAmqpConnection extends ReactorConnection implement
                                            TokenManagerProvider tokenManagerProvider,
                                            MessageSerializer messageSerializer, String product, String clientVersion) {
         super(connectionId, connectionOptions, reactorProvider, handlerProvider, tokenManagerProvider,
-            messageSerializer, product, clientVersion);
+            messageSerializer, product, clientVersion,
+            SenderSettleMode.SETTLED, ReceiverSettleMode.FIRST);
+
         this.connectionId = connectionId;
         this.reactorProvider = reactorProvider;
         this.handlerProvider = handlerProvider;
         this.tokenManagerProvider = tokenManagerProvider;
         this.retryOptions = connectionOptions.getRetry();
         this.messageSerializer = messageSerializer;
-        this.managementChannelMono = getReactorConnection().then(
-            Mono.fromCallable(() -> {
-                return (ServiceBusManagementNode) new ManagementChannel(
-                    createRequestResponseChannel(MANAGEMENT_SESSION_NAME, MANAGEMENT_LINK_NAME, MANAGEMENT_ADDRESS),
-                    connectionOptions.getEntityPath(), connectionOptions.getTokenCredential(),
-                    this.tokenManagerProvider, this.messageSerializer, connectionOptions.getScheduler());
-            }))
-            .cache();
+        this.tokenCredential = connectionOptions.getTokenCredential();
+        this.scheduler = connectionOptions.getScheduler();
+        this.fullyQualifiedNamespace = connectionOptions.getFullyQualifiedNamespace();
+
     }
 
     @Override
-    public Mono<ServiceBusManagementNode> getManagementNode() {
+    public Mono<ServiceBusManagementNode> getManagementNode(String entityPath) {
         if (isDisposed()) {
             return Mono.error(logger.logExceptionAsError(new IllegalStateException(String.format(
-                "connectionId[%s]: Connection is disposed. Cannot get management instance", connectionId))));
+                "connectionId[%s]: Connection is disposed. Cannot get management instance for '%s'",
+                connectionId, entityPath))));
         }
 
-        return managementChannelMono;
+        final ServiceBusManagementNode existing = managementNodes.get(entityPath);
+        if (existing != null) {
+            return Mono.just(existing);
+        }
+
+
+        return getReactorConnection().then(
+            Mono.fromCallable(() -> {
+                final ServiceBusManagementNode node = managementNodes.computeIfAbsent(entityPath, key -> {
+                    final String sessionName = entityPath + "-" + MANAGEMENT_SESSION_NAME;
+                    final String linkName = entityPath + "-" + MANAGEMENT_LINK_NAME;
+                    final String address = entityPath + "/" + MANAGEMENT_ADDRESS;
+
+                    logger.info("Creating management node. entityPath: [{}]. address: [{}]. linkName: [{}]",
+                        entityPath, address, linkName);
+
+                    TokenManager cbsBasedTokenManager =  new AzureTokenManagerProvider(
+                        CbsAuthorizationType.SHARED_ACCESS_SIGNATURE, fullyQualifiedNamespace, entityPath)
+                        .getTokenManager(getClaimsBasedSecurityNode(), entityPath);
+
+                    final Mono<RequestResponseChannel> requestResponseChannel =
+                        createRequestResponseChannel(sessionName, linkName, address);
+                    return new ManagementChannel(requestResponseChannel, messageSerializer, scheduler,
+                        cbsBasedTokenManager);
+                });
+
+                return node;
+            }));
     }
 
     /**